chore(pr): address review comments for #414
- Let malformed annotations fail fast instead of silently catching ParserTypeError
- Improve shape-mismatch error message to include dimension index on inferred side
- Rename test_annotation_dtype_mismatch_on_add to _on_load to match actual test
- Fix stale return type FP16 -> FP32 in test_text_parser complex_op fixture
- Fix pij_out annotation FP32 -> BF16 in system test to match create_tensor dtype

+++ tests/st/codegen/test_paged_attention.py
@@ -151,7 +151,7 @@ class SoftmaxPrepareTestCase(PTOTestCase):
             ]:
                 # Read scale value from config tensor
                 scale: pl.Scalar[pl.FP32] = pl.tensor.read(config, [0])
-                pij_out: pl.Tensor[[16, 128], pl.FP32] = pl.create_tensor([16, 128], dtype=pl.BF16)
+                pij_out: pl.Tensor[[16, 128], pl.BF16] = pl.create_tensor([16, 128], dtype=pl.BF16)
                 mij_out: pl.Tensor[[16, 1], pl.FP32] = pl.create_tensor([16, 1], dtype=pl.FP32)
                 lij_out: pl.Tensor[[16, 1], pl.FP32] = pl.create_tensor([16, 1], dtype=pl.FP32)
                 pij_out, mij_out, lij_out = kernel_softmax_prepare(sij, scale, pij_out, mij_out, lij_out)
